test(#7220): the CSR speedup ratio is stall-discounted, carries the 2x it claimed, and runs in the benchmark lane
compareTraversalPerformance timed two 2-hop traversals with raw System.nanoTime() and asserted their ratio
was greater than 1.0, so a stop-the-world pause charged to the CSR window and not to the OLTP one flipped it
regardless of what either kernel did. The comment above the line claimed "at least 2x"; the code never
encoded that headroom.

Both windows now take the JVM-wide stall out of the span, via StallAwareStopwatch.jvmStallNanos() rather
than a StallAwareStopwatch: the stopwatch reports whole milliseconds and the CSR window is only a couple of
them, so a 1ms quantum on the denominator would leave the ratio dominated by rounding. The bound becomes the
2.0 the comment always claimed - measured idle, CSR wins by 25-40x, so that is a floor with an order of
magnitude of headroom rather than a number tuned to whatever passes today.

The class is tagged benchmark, which is what CLAUDE.md asks for a comparison run, and renamed to end in
Benchmark: the benchmark lane selects on the tag AND -Dsurefire.includes=**/*Benchmark.java, so the tag on
its own would have routed it to no lane at all instead of out of the default one.

The stall discount alone is not what makes this robust, and the comment says so: it recovers the single long
pause #6260 is about, but sustained contention and a series of sub-20ms GC pauses both leave the sampler
ticking and are largely invisible to it. Getting the class out of the 14,539-test lane is the half that
addresses load.
